Callback registration fixes.
Modified Deferred.then() to validate and fail early when non-functions are specified as callbacks.
Modified handling of progress handlers in Deferred.then() to simply update the returned Promise (rather than propagating Promise values or thrown Errors) with either the original progress update or the value returned by the progress handler.