Fixed reported Sencha Touch compatibility issues.
jachenry reported an issue which uncovered a very subtle bug in the targetInstance::initConfig() interception logic.

In two cases in the Deft JS source code, restoring the original method after the first interception of that function is not as simple as anticipated, nor, as it turns out, is it required or desirable behavior.

Fixes #3