Controllable now automatically adds a `getController()` accessor.
If a view is marked with the `Deft.mixin.Controllable` mixin and specifies a `controller` annotation, when the controller class is instantiated, a `getController()` accessor for that controller class instance will automatically be added to the view instance.  When the view is later called to `destroy()`, the `getController()` accessor will be cleared.

Fix #18