Restored support for Object-based `mixins` syntax.
The Sencha class system allows `mixins` to be specified as either an Array or an Object - only the Array-based syntax was working.
Changes in v0.6.6 broke support for the Object-based syntax due to new logic that was introduced that assumed the `mixins` config value was an Array.
Added a Jasmine unit test to detect and avoid any future regression.

Fixes #28