Fixed ViewController cleanup issue reported w/Sencha Touch
Deft.mvc.ViewController::onViewDestroyed() is never called on Sencha Touch (due to the differing destruction lifecycle path). Consequently, the cleanup logic implemented there is not executed on that platform.
Moved the clean-up logic to the destroy() method. This will require any subclasses to call `this.callParent()`, but that should be fairly well understood as a standard practice when overriding destroy() method.
Added associated unit tests to help prevent future regression.

Fixes #45