Added JSDuck comments and units tests for Deft.Logger.  Fixed bugs uncovered in new Deft.Logger unit tests.
The Ext JS log adapter implementation incorrectly attempted to use `is` instead of `=` for assignments, when adapting 'verbose' to 'info' and 'deprecate' to 'warn'.
Ext.Logger has transient availability in Sencha Touch - with a production build it will not be available unless explicitly included.  Revised the Sencha Touch log adapter implementation to accommodate this.
Added the Deft.Logger tests to the Karma configurations and manual test runner.

See #102
See #106